# packages/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: HashHive Control API
  description: |
    Machine-readable REST API for CLI tooling, automation platforms, and CI
    integrations. Authenticated by per-user API keys (format `cst_*`),
    paginated with `offset`/`limit`, and emits RFC 9457 problem-details on
    errors.
  version: 1.0.0
servers:
  - url: /api/v1/control
    description: Control API base path
paths:
...
  /hashlists:
    get:
      operationId: listHashLists
      summary: List hash lists in the active project
      tags: [HashLists]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated hash lists
        "400":
          $ref: "#/components/responses/ValidationError"
  /hashlists/{id}:
    get:
      operationId: getHashList
      summary: Get a hash list
      tags: [HashLists]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Hash list record
        "404":
          $ref: "#/components/responses/NotFound"
  /hashlists/{id}/stats:
    get:
      operationId: getHashListStats
      summary: Crack statistics for a hash list
      tags: [HashLists]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Stats payload
        "404":
          $ref: "#/components/responses/NotFound"
...
  /campaigns:
    get:
      operationId: listCampaigns
      summary: List campaigns
      tags: [Campaigns]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, running, paused, completed, cancelled]
      responses:
        "200":
          description: Paginated campaigns
    post:
      operationId: createCampaign
      summary: Create a campaign
      tags: [Campaigns]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, hashListId]
              properties:
                name:
                  type: string
                description:
                  type: string
                hashListId:
                  type: integer
                priority:
                  type: integer
      responses:
        "201":
          description: Created campaign
        "400":
          $ref: "#/components/responses/ValidationError"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      summary: Get a campaign
      tags: [Campaigns]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Campaign record
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCampaign
      summary: Update a campaign
      tags: [Campaigns]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                priority:
                  type: integer
      responses:
        "200":
          description: Updated campaign
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/transition:
    post:
      operationId: transitionCampaign
      summary: Transition a campaign to a new status
      tags: [Campaigns]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetStatus]
              properties:
                targetStatus:
                  type: string
                  enum: [draft, running, paused, completed, cancelled]
      responses:
        "200":
          description: Updated campaign
        "409":
          $ref: "#/components/responses/Conflict"
  /attacks:
    get:
      operationId: listAttacks
      summary: List attacks within a campaign
      tags: [Attacks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - in: query
          name: campaignId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of attacks
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createAttack
      summary: Create an attack on a campaign
      tags: [Attacks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [campaignId, mode]
              properties:
                campaignId:
                  type: integer
                mode:
                  type: integer
                hashTypeId:
                  type: integer
                wordlistId:
                  type: integer
                rulelistId:
                  type: integer
                masklistId:
                  type: integer
      responses:
        "201":
          description: Created attack
  /attacks/{id}:
    get:
      operationId: getAttack
      summary: Get an attack
      tags: [Attacks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Attack record
    patch:
      operationId: updateAttack
      summary: Update an attack
      tags: [Attacks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Updated attack
    delete:
      operationId: deleteAttack
      summary: Delete an attack
      tags: [Attacks]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "204":
          description: Deleted
  /agents:
    get:
      operationId: listAgents
      summary: List agents in the active project
      tags: [Agents]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: status
          schema:
            type: string
            enum: [online, offline, busy, error, benchmarked]
      responses:
        "200":
          description: Paginated agents
  /agents/{id}:
    get:
      operationId: getAgent
      summary: Get an agent
      tags: [Agents]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Agent record
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAgent
      summary: Update an agent (admin only)
      tags: [Agents]
      security:
        - BearerApiKey: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/ProjectIdHeader"
      responses:
        "200":
          description: Updated agent
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: cst_<userId>_<random>
  parameters:
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        minimum: 1
    ProjectIdHeader:
      in: header
      name: X-Project-Id
      description: Active project the request operates against
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details envelope
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path:
                type: string
              code:
                type: string
              message:
                type: string
...
  responses:
...
    Forbidden:
      description: Caller lacks permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Target resource was not found or is not visible to the caller
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request body or query parameters failed validation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Operation conflicts with current resource state
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"